requirements.txt
pandas
numpy
matplotlib
mlflow
pyaml
uvicorn
scikit-learn
fastapi
pymongo
certifi
python-dotenv